+++ src/main/java/GameBoard/Main.java
@@ -1,5 +1,7 @@
 package GameBoard;
 
+import at.aau.serg.websocketdemoserver.GameObjects.Player;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
