+++ src/main/java/at/aau/serg/websocketdemoserver/GameManager/GameManager.java
@@ -8,7 +8,6 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Random;
 import java.util.UUID;
